Chake::Backend: run commands by opening a shell and writing to it
One of the actions I perform a lot is running `pgrep something` on
servers I manage. pgrep never returns itself, but it does return the
shell running it, and the ssh command running that shell. To avoid that,
we need to not have any process with "something" in its command line.

This also opens up the possibility of running an entire script on the
nodes.