converge, apply: allow removing data from the node JSON attributes
The -j/--json-attributes option only adds data; so if we deleted some
entries locally, when applying they wouldn't be deleted on the server
side. By wiping the existing json file before running chef-solo, it will
get recreated and will only contain the data that is currently in the
node definition.