Implement new configuration manager: itamae-remote
This configuration manager is similar to the itamae one, but instead of
running itamae on the workstatation running chake, it runs itamae
remotely on each node (like the Chef one runs chef-solo remotely on each
node). In my tests, this provides a speed up of an order of magnitude
when converging a previously converged virtual machine with a reasonably
large configuration.